Make view annotation manager class instead of an interface
